Fixed convey.async/<<< function so that channel get closed right after callback call.
This is a common behaviour that callback function get called only once. While it would be useful to support calling callback multiple times, but in this case we gonna leak - there is no way to determine when the channel can be closed. More over besides leaking there would be a deadlock whenever somebody decide to read all the values from the channel